distance_prices

+++ database/migrations/2024_01_23_200244_create_distance_prices_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->decimal('min_km', 5, 2);
             $table->decimal('max_km', 5, 2);
-            $table->unsignedBigInteger('load_type_id');
+            $table->integer('load_type_id');
             $table->integer('price');
             $table->timestamps();
